Revert "Merge pull request #1256 from agcomphr25/codex/p1-production-queue-po-orders"
This reverts commit 0dfc6a27a99e6e95dca28b65e78c22353e684264, reversing
changes made to b11ef68866c5da39571aa421da3913cce15bba16.
